refactor: unify config files
