+++ src/components/sections/Skills.astro
@@ -16,8 +16,7 @@ import SwiftUI from "@/icons/swiftui.astro";
 import Kotlin from "@/icons/kotlin.astro";
 import Flutter from "@/icons/flutter.astro";
 
-import { setLanguage } from "@/i18n/translations";
-import { getCV } from "@/i18n/supportedLangs";
+import { getCV, setLanguage } from "@/i18n/config";
 
 const SKILLS_ICONS: Record<string, any> = {
   HTML,
