+++ src/i18n/LangPicker.astro
@@ -1,5 +1,5 @@
 ---
-import { getLangs } from "@/i18n/supportedLangs";
+import { getLangs } from "@/i18n/config";
 import { getRelativeLocaleUrl } from "astro:i18n";
 
 const langs = getLangs();
