+++ src/layouts/Layout.astro
@@ -1,5 +1,5 @@
 ---
-import { getCV } from "@/i18n/supportedLangs";
+import { getCV } from "@/i18n/config";
 
 interface Props {
   title: string;
