+++ src/pages/index.astro
@@ -9,8 +9,7 @@ import Projects from "@/components/sections/Projects.astro";
 import Skills from "@/components/sections/Skills.astro";
 import KeyboardManager from "@/components/KeyboardManager.astro";
 
-import { setLanguage } from "@/i18n/translations";
-import { getCV } from "@/i18n/supportedLangs";
+import { setLanguage, getCV } from "@/i18n/config";
 import LangPicker from "@/i18n/LangPicker.astro";
 
 const locale = Astro.currentLocale;
